First Draft UI
First 'draft' of the UI is now somewhat complete. No major breakthroughs
but still getting used to using GitHub and Unity (hehe).

Done:
Button icons added.
Panel implemented to group the various clothing buttons together,
situated at the top of the window.

To do:
Unfortunatley most of the buttons still aren't functional, thinking
about implementing a grid overlay that comes up on click for users to
select the different type of shirt etc. Will be working on this next.
Need icon for hand props?
Tooltips?
Some of the spacing, styling etc. Might need some work unless everyone
is happy with the way it is.